Started new mipmaps and masks folder structure system that limits number of files to 100 per folder by using the following file and folder naming scheme:
Patch id = 12345  --->  <level>/12/34/5.filename.jpg

For masks the same without the <level>.

Also and VERY IMPORTANT, the trakem2.mipmaps and trakem2.masks
are now stored under a single
trakem2.<timestamp>.<folder-hash>.<username-hash> parent folder, that
exists in the user-chosen storage folder. This folder and the xml file
are all that is required for a TrakEM2 project to exist on the
filesystem.