Fixed alpha mask painting problem.
What was the problem: for some reason, the Area.intersect method
generates different results if a new Area(new Rectangle(...)) is
intersected with the Area extracted from the ShapeRoi, than the other
way around. It only gave problems when the x coordinate was negative.
By setting the Area(Rectangle) to the intersection of itself and the
Area from the ShapeRoi, then there never is a negative coordinate.