Fixed file path repair support for stacks, and enhanced usability.
Now the file path repair window an be closed at user's choice.
If the file name to fix a path with doesn't match, it complains and asks
for confirmation, or at least prints a warning.

Also renamed Loader.getAbsoluteFilePath to getImageFilePath, for
consistency with Patch method.

Thank you to Jean-Yves Tinevez for testing and reporting this bug.