Keep Points in world space and correct for affine invariance of alpha by chaining the transforms.  The actual MLST runs in world space as expected by the user who edits it.
Minor speed up by less excessive object generation.  It seems that
creating Objects in Java is cheap but definitely not for free.

TODO
Apply is not yet done at all.
mousePressed uses a TransformMesh of wrong size and location such that
depending on magnification, the location cannot be inverted.  I know how
to fix but have to sleep and travel first.

BTW, the update of ImageJA incl. precompiled/ij.jar fixes the unintended
white background issue as expected.