Revert "Revert "An (as yet largely untested) patch to let TrakEM2 build with openjdk""
This reverts commit e702e2700c269e81e20b492853a36ce704cd15c8.

The problem was a misunderstanding: the patch _was_ largely untested.  As
none of the primary TrakEM2 authors had time to test this patch, which
was provided already over a month ago, and which is blocking our Debian
release of Fiji Plzeň, yours truly just tested the exact recommended
procedure mentioned in bug 53.

So while it _is_ largely untested at this point, it is _not_ untested,
but tested in _exactly_ the requested ways.