added correspondence weight as a parameter to Align.Param
This weight makes mostly sense do distinguish between the
influences of intra- vs. cross-layer correspondences in the
multi-layer-mosaicking.  Higher weights for cross-section
correspondences will result in better continuity across sections
to the cost of more deformation in the section.  Keep in mind that the
global amount of deformation is minimized anyway.