added option for alternative color cues
When lots of treelines are in one spot, and color cues are on for sections above & below the current section, the world turns into a sea of red & blue wires. This option is a quick look at what happens if instead wires are drawn using the treeline color as a base, with the wires getting darker or ligher going away from the current section. It seems to help visual parsability somewhat.