Removed recreation of Buckets from ProjectTree.remove, which was there as a patch to an earlier error, and also added a remove(boolean,Set<Displayable>) that optimally finds the ProjectThing instances and nodes wrapping each Displayable.
For the purpose, a new Displayable.softRemove has been added,
which calls Displayable.remove when not overriden. It is intended for ZDisplayable
and clears only the contents, but doesn't alter the object membership in trees
or container lists like the ZDisplayable list of the LayerSet.

Selection now calls this new ProjectTree.remove method.