Implemented new cache for mipmap java.awt.Image. A comparison with the old CacheImageMipMaps:
Old insert 500000: ~4 seconds
Old remove 10 x 50000: ~0.4 seconds

New insert 500000: ~4 seconds
New remove 10 x 50000: ~0.4 seconds

BUT:

The old, which was using a map and a FIFO LinkedList,
was NOT an access-time based priority list.
And the old had linear scalability problems beyond a few thousand images.

The new is a priority list, chunked at groups of 20 ids, each id with all its mipmaps.
When an image for a given id is accessed, all images for that id are put into
the highest priority for preservation.
And the new does not have scaling problems.

At low number of images, both perform just as well, with the new having
the advantage of actually preserving recently accessed images.