Ensure clipRect is always the whole area of the canvas when triggering a component repaint on the canvas.
This fixes the error of lack of canvas refresh, evident
as low-res images still painted in the canvas.

What hinted me as to the error was that moving
the canvas window beyond the desktop boundaries
refreshed the parts of the canvas that went beyond
the screen but not the still-blurred parts still
within the screen.