Fix for a race condition when starting painting
It was possible for the last mouseMoved event (with no buttons
held down) to update flags after the mousePressed and mouseDragged
events (with the left button held down) had set them, which would
cause the Painter thread to exit immediately.  This can be fixed
by updating the flags in mouseMoved() before dispatching the event
to the MouseMovedThread for further processing.