AreaList measurements now include the X,Y,Z of the center of mass.
Also fixed errors in the measurements where the Z and thickness
was always for a fixed this.layer (from parent class Displayable)
and not from the actual layer where an area was being measured.
In practice only the measurement of the maximum diameter is affected,
given that the Z was always zero (the first layer is the default).