FSLoader.generateMipMaps pushes forward the creation of BufferedImages, which for the raw format are not necessary. The IntegralImageMipMaps has been adapted to deliver such barebones images as well, packed into an ImageBytes class.
Additionally, the ImageSaver got methods to create TYPE_INT_RGBA_PRE images
directly from an int[] array where alpha is premultiplied, by using the
appropriate DirectColorModel. This opened the possibility of avoiding
duplicating RAM needs just to create a TYPE_INT_ARGB_PRE BufferedImage
when loading raw images.

Currently the premultiplication of alpha is done when loading the mipmap
rather than when saving it. This could be easily fixed, if the ImageIO
turns out to be able to correctly load images and recognize them as premultiplied.