The CoordinateTransform of the Patch is now always paged out to folder trakem2.cts/. The paging occurs when parsing an XML file, and also when setting a new CoordinateTransform to the Patch.
Each Patch now gets a unique ID for the CoordinateTransform,
which is used to generate a unique file name.

Old files are not deleted until running the explicit command
to delete stale files under "Project - Delete stale files...".
This is so because there is no easy way of determing whether
some files are really stale or not: other XML files, resulting
from copies in the file system or copies from "Save as...", may
still point to what are old coordinate transform files for the
currently loaded XML file.