patch does not propagate NonInvertibleModelException from TransformMesh which is merely a hint that coordinates outisde hte mesh have been transferred by extrapolation
patch got getLocalAffine(x,y) method which delivers an affine transform
for a specific location.  This affine transform is either the global
affine transform of the patch or the combined affine transform of the
local affine transform in the transform mesh that approximates the
non-linear transformation of the patch and its global affine transform.