threadsafe Utils.ensure(File)
There is a race condition in File.mkdirs() that makes it fail if another
thread attempts to generate the exact same directory simultaneously.
This does not happen frequently and thus remained undetected for a long
time, but it reliably screws up large and complicated alignments on beefy
workstations.  Bug reference

http://bugs.java.com/view_bug.do?bug_id=6492293