Include ImgLib1 wholesale
TrakEM2 is the sole remaining user of the now thoroughly obsolete ImgLib1.
Rather than pretend that ImgLib1 is a supported, maintained software
library, let's just put it into TrakEM2, with the intention to move on to
ImgLib2 as time allows.

There was a brief consideration whether the ImgLib1 classes should be
shaded or not (i.e. moved into a different Java package to avoid clashing
with other ImgLib1 versions on the class path). However, TrakEM2
encouraged the development of TrakEM2 plugins and any of them are
potential users of ImgLib1' data structures (and thus, backwards
compatibility would be broken by shading ImgLib1 in TrakEM2).

The included code comes verbatim from ImgLib1's f96f8c84a(Create
README.md) – and adds JUnit as test scope dependency as it is now
needed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>