POM: add enforcer configuration to fix the build
In the case of requireReleaseDeps, the new configuration simply excludes
the TrakEM2 artifacts themselves, since they are part of the reactor and
should not be checked. (The fact that they are checked is a limitation
of the rule; see https://jira.codehaus.org/browse/MENFORCER-185). An
unfortunate but minor side effect is that the enforcer will now not flag
the case where pom-trakem2 extends a SNAPSHOT version of pom-fiji, but
such a situation is unlikely to occur by accident.

A stickier problem is the configuration of requireReproducibleBuilds,
the extra rule provided by the scijava-maven-plugin. This rule does a
more thorough check than requireReleaseDeps does, and unfortunately at
the moment cannot be disabled at the level needed to make the build pass
for multi-module projects like this one. So as a hack, this commit
temporarily disables the enforcer completely. As soon as we have a more
configurable version of the requireReproducibleBuilds rule, we should
reenable the enforcer here, and configure it accordingly to avoid any
failures due to the SNAPSHOT couplings within the multi-module build.