added flag and checkbox to consider only the widest set of landmarks for series alignment
This is useful in situations where sections series are acquired that show
parts of previous and subsequent sections together with the current
section (happens e.g. with tightly trimmed blocks in TEM).  The flag then
assures that of the several possible hypotheses to align such an image in
the series, only that supported by the point set with largest spatial
extent is considered.  This is usually the section of interest.