Loader.getFlatAWTImage is now safer.
First, AWT images beyond 1 GB are used less. Before, they could
be created when the the "quality" flag was set to true, or when
there weren't any mipmaps which used to set the scale to 1.0.

Images larger than 2 GB will continue to fail with a NegativeArraySizeException.

Second, when "quality" is true, an effort is made to use the largest
image possible (before it was double the required scale, up to a maximum
 of 1.0), with the limit at 1 GB. While 2 GB BufferedImage are possible,
these often have problems: noise, parts not painted, etc that don't
throw any errors. We consider only BufferedImage instances at the safer
limit of 1 GB.