Patch.makeFlatGreyImage: the safest shot at getting the correct image.
Will use AWT and mipmaps if available and the target image is under 1 GB.
Otherwise will use mipmaps if available to paint using ExportUnsignedByte.
Othrewise will use ExportUnsignedShort to the largest image possible
that is still under 2 GB, and then Gaussian-downsample to the requested size.

When using mipmaps there is no need to downsample, as mipmaps are already
properly downsampled when they are created.